feat: add render-pictures
