+++ index.html
@@ -162,7 +162,7 @@
             </li>
             <li class="social__comment">
               <img class="social__picture" src="img/avatar-3.svg" alt="Аватар комментатора фотографии" width="35" height="35">
-               <p class="social__text">Да это фоташоп!!!!!!!!</p>
+                <p class="social__text">Да это фоташоп!!!!!!!!</p>
             </li>
           </ul>
