+++ js/main.js
@@ -1,9 +1,6 @@
-import {makePhoto} from './generate-photos.js';
+import {makePhoto} from './generate-posts.js';
+import {createPosts} from './render-pictures';
+const generatePosts = Array.from({length: 25}, makePhoto);
 
-const generatePhotos = Array.from({length: 25}, makePhoto);
-
-console.log(generatePhotos);
-/*
-если вызвать просто в консоли, не работает -> Uncaught ReferenceError: generatePhotos is not defined
-at <anonymous>:1:13
-*/
+console.log(createPosts(generatePosts));
+//export {generatePosts};
